requirements.txt
streamlit
opencv-python-headless
numpy
Pillow
torch
torchvision
openai